back-end/requirements.txt
asgiref==3.8.1
black==25.1.0
boto3==1.38.23
botocore==1.38.23
click==8.2.1
Django==5.2.1
django-configurations==2.5.1
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
gunicorn==23.0.0
jmespath==1.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
